web: add absolute relative path
