--- web/massastation/src/utils/massaConstants.ts
@@ -1,12 +1,0 @@
-// We store each possible variation of the massas's name in the array
-export const massalabsNomination = ['Massa Labs'];
-// Constants to mimic server status
-
-export const PLUGIN_START = 'start';
-export const PLUGIN_STOP = 'stop';
-export const PLUGIN_UPDATE = 'update';
-
-export const isZipFile = /^(http)[^\s]*\.zip$/i;
-
-// List of websites that are allowed to be added to favorites
-export const FAVORITES_WEBSITES = ['flappy', 'psychedelic', 'flappynathana1'];
