+++ web/massastation/vite.config.ts
@@ -1,6 +1,7 @@
 import { defineConfig, loadEnv } from 'vite';
 import react from '@vitejs/plugin-react';
 import svgr from 'vite-plugin-svgr';
+import * as path from 'path';
 
 export default ({ mode }) => {
   // loadEnv(mode, process.cwd()) will load the .env files depending on the mode
@@ -16,6 +17,12 @@ export default ({ mode }) => {
       manifest: true,
       sourcemap: true,
       assetsDir: './', // put the assets next to the index.html file
+    },
+    resolve: {
+      alias: [
+        { find: '@', replacement: path.resolve(__dirname, 'src') },
+        { find: '@/shared', replacement: path.resolve(__dirname, '../shared') },
+      ],
     },
   });
 };
