+++ web/shared/consts/index.ts
@@ -1,0 +1,1 @@
+export * from "./theme";
