+++ web/shared/index.ts
@@ -1,0 +1,2 @@
+export * from "./consts";
+export * from "./interfaces";
